Break parts out of core.ts
