+++ src/compiler/utilities.ts
@@ -2,9 +2,6 @@
 
 /* @internal */
 namespace ts {
-    export const emptyArray: never[] = [] as never[];
-    export const emptyMap: ReadonlyMap<never> = createMap<never>();
-
     export const externalHelpersModuleNameText = "tslib";
 
     export interface ReferencePathMatchResult {
