+++ tests/baselines/reference/api/tsserverlibrary.d.ts
@@ -2703,12 +2703,12 @@ declare namespace ts {
     }
 }
 declare namespace ts {
+    function isExternalModuleNameRelative(moduleName: string): boolean;
+}
+declare namespace ts {
     const versionMajorMinor = "2.7";
     /** The version of the TypeScript compiler release */
     const version: string;
-}
-declare namespace ts {
-    function isExternalModuleNameRelative(moduleName: string): boolean;
 }
 declare function setTimeout(handler: (...args: any[]) => void, timeout: number): any;
 declare function clearTimeout(handle: any): void;
